feat: add pagination component for cube exploration page
- Introduced a new Pagination component to handle page navigation, improving user experience.
- Removed inline pagination logic from the cube exploration page and replaced it with the new component for better code organization and maintainability.
- Updated the layout to ensure responsive design and accessibility features for pagination buttons.
